backend/requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
sqlmodel==0.0.21
python-multipart==0.0.9
PyJWT==2.9.0
google-auth==2.33.0
google-auth-oauthlib==1.2.1
google-auth-httplib2==0.2.0
python-dotenv==1.0.1
email-validator==2.2.0
Pillow==10.4.0
requests==2.31.0
mariadb>=1.1,<2
pymysql>=1.0,<2